Setup `gaussian_filter_nan` in utils, use in `estimate_correlation_from_phase` (#420)
* Setup `gaussian_filter_nan` in utils, use in `estimate_correlation_from_phase`

* Use Gaussian in `interferogram`
